% --- Source file: Bullying.Rd --- 
\name{Bullying} 
\alias{Bullying} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
punteggi di un campione di bambini su una scala di percezione della pressione subita dai coetanei \code{PPP} e una scala di comportamento passivo.} 

\usage{data( Bullying )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{class}: Factor, classe di appartenenza.
    \item \code{age}: num, età.
    \item \code{PPP}: num, Perceived Peer Pressure.
    \item \code{PBB}: num, Passive bystanding behavior.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
