% --- Source file: crimi.Rd --- 
\name{crimi} 
\alias{crimi} 
\title{Data set fittizio} 
\description{ 
Data set fittizio: studio sulla paura della criminalità e le sue determinanti, raccolti in 4 diversi quartieri di una grande città.
} 

\usage{data( crimi )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{Preg}: num, grado di pregiudizio verso extracomunitari.
    \item \code{Paura}: num, paura della criminalità.
    \item \code{Strat}: num, strategie per fronteggiare la paura. 
    \item \code{Extra}: num, percezione della presenza di extracomunitari.
    \item \code{DisFis}: num, percezione di disordine fisico. 
    \item \code{DisSoc}: num, percezione di disordine sociale.
    \item \code{Vitt}: num, grado di vittimizzazione.
    \item \code{Quartiere}: Factor, quartiere di residenza.
    }
}

\references{Santinello, M., Vieno, A., Davoli, K., Pastore, M. (2005). Il modello contesto-coping-adattamento per la spiegazione della paura della criminalità . \emph{Giornale Italiano di Psicologia}, 32, 161-180.}
\author{Massimiliano Pastore}
\keyword{data}
