% --- Source file: donne.Rd ---
\name{donne}
\alias{donne}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}

\description{
Data set fittizio: risposte ad un questionario di atteggiamento sull'attivita'
extradomestica in un campione di 81 donne con figli e non. 
}

\usage{
data( donne )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{figli}: Factor, indica se la donna ha figli (\code{si}) oppure no.
    \item \code{attegg}: num, punteggi di atteggiamento. 
        Piu' e' alto il punteggio maggiore il grado di favore verso l'attivita' extradomestica.}
}

%- maybe also 'usage' for other objects documented here.

%\value{
%Contiene le frequenze (fittizie) di un campione di 210 donne 
%di cui sono rilevate livello d'ansia e tempo dedicato alla preparazione di un esame di statistica.
%}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
