% --- Source file: earlymath.Rd ---
\name{earlymath}
\alias{earlymath}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set}

\description{
Data set adattato da Passolunghi & al. 2014: Dati di una ricerca sulle capacità matematiche nei bambini di 6 anni.
}

\usage{
data( earlymath )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{gender}: Factor, sesso.
    \item \code{MAT}: num, misura delle abilità matematiche. 
    \item \code{QI}: num, livello di intelligenza. 
    \item \code{WM}: num, memoria di lavoro.
    \item \code{STM}: num, memoria a breve termine.
    \item \code{ANS}: num, misura la capacità che permette di stimare ad esempio i risultati di operazioni per via approssimata (\emph{Approximate Number Sistem}).
    }
}
\source{Passolunghi, M.C., Cargnelutti, E., Pastore, M. (2014). The contribution of general cognitive abilities and approximate number system to early mathematics. \emph{British Journal of Educational Psychology}, 84, 631-649.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
