% --- Source file: ecoansia.Rd ---
\name{ecoansia}
\alias{ecoansia}
\alias{legenda}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set}

\description{
Adattato da Lutz & al. 2023: Dati di una ricerca sull'ecoansia. 
}

\usage{
data( ecoansia )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{numid}: Factor, codice identificativo
    \item \code{dayNum}: num, giorno di rilevazione.
    \item \code{age}: num, età.
    \item \code{sex}: Factor, sesso.
    \item \code{EA}: num, ecoansia. 
    \item \code{PAA}: num, Positive Activated Affect. 
    \item \code{PDA}: num, Positive Deactivated Affect.
    \item \code{NAA}: num, Negative Activated Affect.
    \item \code{NDA}: num, Negative Dectivated Affect.
    \item \code{Coh}: num, Coherence.
    \item \code{GPEB}: num, General Pro-environment.
    \item \code{SPEB}: num, Specific Pro-environment.
    }
}
\references{Lutz, P. K., Zelenski, J. M., & Newman, D. B. (2023). Eco-anxiety in daily life: Relationships with well-being and pro-environmental behavior. \emph{Current Research in Ecological and Social Psychology}, 4, 100110.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
