% --- Source file: elektrosmoke.Rd ---
\name{electrosmoke}
\alias{electrosmoke}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}

\description{
Data set fittizio: Dati di una ricerca sull'utilizzo della sigaretta elettronica. 
}

\usage{
data( electrosmoke )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{age}: num, anni dei soggetti.
    \item \code{smokeyears}: num, numero di anni da fumatore.
    \item \code{MQS}: num, punteggio di motivazione a smettere di fumare.
    \item \code{gender}: Factor, genere del soggetto.
    \item \code{cigday}: num, sigarette fumate al giorno.
    \item \code{elcig}: Factor, variabile dicotomica che esprime se il soggetto ha provato (\code{yes}) o no la sigaretta elettronica.    }
}
%\source{Tabacknick, B.G., Fidell, L. (1996), \emph{Using Multivariate Statistics}, Harper Collins, NY.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
