% --- Source file: ESdata.Rd ---
\name{ESdata}
\alias{ESdata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set}

\description{
Data set adattato da Lionetti & al. 2019: Dati di una ricerca sulla \emph{Environmental Sensitivity} in un campione di 200 bambini. 
}

\usage{
data( ESdata )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{ParSty}: num, misura dello stile genitoriale permissivo.
    \item \code{HighSens}: Factor, livello di sensibilità \code{1 = high}.
    \item \code{ChildBeh}: num, misura di problemi comportamentali.
    \item \code{gender}: Factor, sesso.
    \item \code{age}: num, età.
    }
}
\references{Lionetti, F., Aron, E. N., Aron, A., Klein, D. N.,  Pluess, M. (2019). Observer-rated environmental sensitivity moderates children’s response to parenting quality in early childhood. \emph{Developmental psychology}, 55, 2389.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
