% --- Source file: firstaid.Rd --- 
\name{firstaid} 
\alias{firstaid} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
punteggi di attitudine al pronto soccorso in un campione di 562 soggetti.} 

\usage{data( firstaid )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{attitude}: num, punteggi di attitudine al pronto soccorso.
    \item \code{sex}: Factor, sesso dei soggetti.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
