% --- Source file: Gini.Rd ---
\name{Gini}
\alias{Gini}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set: Indice di Gini e speranza di vita}
\description{
Data set: Indice di Gini e speranza di vita nelle regioni italiane.
}
\usage{
data( Gini )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{regione}: chr, denominazione regione.
    \item \code{gini}: num, indice di Gini (1995-2000).
    \item \code{life}: num, aspettativa di vita alla nascita (2001).
    \item \code{residenti}: num, popolazione residente (2018).
    \item \code{n.comuni}: num, numero di comuni.
    \item \code{n.provincie}: num, numero di provincie.
    }
}
%- maybe also 'usage' for other objects documented here.

%\value{
%Contiene le frequenze (fittizie) di un campione di 210 Gini 
%di cui sono rilevate livello d'ansia e tempo dedicato alla preparazione di un esame di statistica.
%}
\author{Massimiliano Pastore}
\keyword{data}
