% --- Source file: impiegati.Rd --- 
\name{impiegati} 
\alias{impiegati} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
dati di un campione di impiegati di banca.} 

\usage{data( impiegati )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{id}: num, codice dell'impiegato.
    \item \code{sesso}: Factor, Sesso.
    \item \code{istruz}: Factor, Anni scolastici - livello di istruzione.
    \item \code{catlav}: Factor, Categoria lavorativa.
    \item \code{stipatt}: num, Stipendio attuale annuo.
    \item \code{stiiniz}: num, Stipendio iniziale annuo.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
