% --- Source file: inibition.Rd --- 
\name{inibition} 
\alias{inibition} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
dati di una ricerca sui meccanismi di inibizione cognitiva.} 

\usage{data( inibition )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{Code}: Factor, codice soggetto.
    \item \code{Age}: num, età.
    \item \code{Ageclass}: Factor, età in classi.
    \item \code{INI}: num, punteggio di inibizione.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
