% --- Source file: kidiq.Rd ---
\name{kidiq}
\alias{kidiq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}

\description{
Data set fittizio: 434 coppie madri figli. 
}

\usage{
data( kidiq )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{kid_score}: int, punteggio del figlio ad un test cognitivo.
    \item \code{mom_hs}: Factor, livello di istruzione della madre: \code{1} = diplomata.
    \item \code{mom_iq}: num, punteggio della madre al QI.
    \item \code{mom_work}: Factor, tipo di lavoro della madre.
    \item \code{mom_age}: num, età della madre.}
}
\source{Fox, J. (1996). \emph{Applied Regression Analysis, Linear Models, and Related Methods}. Harper Collins, NY.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
