% --- Source file: kluegel.Rd ---
\name{kluegel}
\alias{kluegel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set}

\description{
Data set adattato da Kluegel, Singleton & Starnes (1977). 
}

\usage{
data( kluegel )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{reddperc}: num, punteggio di reddito percepito.
    \item \code{presperc}: num, punteggio di prestigio percepito.
    \item \code{statperc}: num, punteggio di status percepito.
    \item \code{reddreal}: num, punteggio di reddito reale.
    \item \code{presreal}: num, punteggio di prestigio reale. 
    }
}
\references{Kluegel, J. R., Singleton Jr, R., & Starnes, C. E. (1977). Subjective class identification: A multiple indicator approach. \emph{American Sociological Review}, 599-611.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
