% --- Source file: lesi.Rd ---
\name{lesi}
\alias{lesi}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}
\description{
Data set fittizio: punteggi di 50 soggetti con lesioni cerebrali ad
un test per la valutazione delle funzioni cognitive e relativo follow up.
}
\usage{
data( lesi )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{test}: num, punteggio al test nella prima somministrazione.
    \item \code{follup}: num, punteggio al test nella seconda somministrazione.}

I punteggi sono espressi in modo che valori bassi indicano una maggiore compromissione 
delle funzioni cognitive.
}
%- maybe also 'usage' for other objects documented here.

%\value{
%Contiene le frequenze (fittizie) di un campione di 210 lesi 
%di cui sono rilevate livello d'ansia e tempo dedicato alla preparazione di un esame di statistica.
%}
\author{Massimiliano Pastore}
\keyword{data}
