% --- Source file: lettura.Rd --- 
\name{lettura} 
\alias{lettura} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
punteggi su un compito di lettura, relativi ad un campione di bambini 
raccolti in diverse scuole.} 

\usage{data( lettura )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{id}: Factor, codice soggetto.
    \item \code{scuola}: Factor, codice scuola.
    \item \code{reddito}: num, reddito annuale della famiglia in euro.
    \item \code{punteggio}: num, punteggio sul compito di lettura.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
