% --- Source file: mathieu.Rd ---
\name{mathieu}
\alias{mathieu}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}

\description{
Data set fittizio: Dati di una ricerca sui predittori della soddisfazione e dell’impegno organizzativo presso i cadetti impegnati in corsi di addestramento militare.
}

\usage{
data( mathieu )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{Role.strain}: num, percezione di stress/tensione associata all’occupare un determinato ruolo.
    \item \code{Training.char}: num, caratteristiche della formazione ricevuta.
    \item \code{Satisfaction}: num, soddisfazione percepita.
    \item \code{Organ.commit}: num, percezione dell’impegno dell’unità organizzativa.
    \item \code{Unit.cohesion}: num, coesione della classe in cui vengono svolti i corsi.
    \item \code{Vet.status}: num, esperienza in ambito militare.    
    \item \code{Unit.perf.stand}: num, percezione individuale di quanto la propria unità militare sia stata adeguatamente formata per il raggiungimento di obiettivi elevati 
    \item \code{Achiev.mot}: num, motivazione al raggiungimento di obiettivi.
    }
}
%\source{Tabacknick, B.G., Fidell, L. (1996), \emph{Using Multivariate Statistics}, Harper Collins, NY.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
