% --- Source file: mathschool.Rd ---
\name{mathschool}
\alias{mathschool}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}

\description{
Data set fittizio: punteggi di fine anno al test di matematica rilevati in 7 classi di una scuola nella successione dei 5 anni scolastici utilizzati per valutare l'apprendimento della matematica. 
}

\usage{
data( mathschool )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{subj}: Factor, codice studente.
    \item \code{classe}: Factor, codice classe.
    \item \code{docente}: Factor, codice docente.
    \item \code{anno}: num, anno di corso.
    \item \code{math}: num, voto in matematica.
    }
}
%%\source{Tabacknick, B.G., Fidell, L. (1996), \emph{Using Multivariate Statistics}, Harper Collins, NY.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
