% --- Source file: memoria.Rd --- 
\name{memoria} 
\alias{memoria} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: Relazione tra intelligenza e memoria.} 

\usage{data( memoria )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{QI}: num, punteggio di intelligenza.
    \item \code{memoria}: num, punteggio di memoria, rilevato dopo una tecnica di addestramento.
    \item \code{tecnica}: Factor, tecnica di miglioramento della memoria utilizzata.
    \item \code{sogg}: num, codice soggetto.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
