% --- Source file: MVM.Rd ---
\name{MVM}
\alias{MVM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}
\description{
Data set fittizio per esempi di modelli multivariati.
}
\usage{
data( MVM )
}
%- maybe also 'usage' for other objects documented here.

\format{
Il data frame contiene le seguenti variabili:
\itemize{
    \item \code{Y1}: num.
    \item \code{Y2}: num.
    \item \code{Y3}: num.
    \item \code{X1}: num. }
}

\author{Massimiliano Pastore}
\keyword{data}
