% --- Source file: OBQ.Rd --- 
\name{OBQ} 
\alias{OBQ} 
\title{Data set fittizio} 
\description{ 
Data set 
fittizio: campione di 51 pazienti con disturbi ossessivi, seguiti con una terapia in uno studio longitudinale per 5 anni. Per misurare il livello di disturbo viene utilizzato il test OBQ (Obsessive Beliefs Questionnaire) in cui valori alti indicano livelli di ossessione alti. 
} 

\usage{data( OBQ )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{time}: num, anno di raccolta (0 = baseline).
    \item \code{subj}: Factor, codice paziente.
    \item \code{values}: num, punteggi OBQ.}
}

\references{Novara, C., Pastore, M., Ghisi, M., Sica, C., Sanavio, E., McKay, D. (2011). Longitudinal aspects of obsessive compulsive cognitions in a non-clinical sample: A five-year follow-up study. \emph{Journal of Behavior Therapy and Experimental Psychiatry}, 42, 317-324.}
\author{Massimiliano Pastore}
\keyword{data}
