% --- Source file: Omega.Rd --- 
\name{Omega} 
\alias{Omega} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
popolazione di voti ad un esame.} 

\usage{data(Omega)} 

%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}

\examples{
data(Omega)
barplot(table(Omega))
}
\keyword{data}
