% --- Source file: parenting.Rd --- 
\name{parenting} 
\alias{parenting} 
\title{Data set fittizio} 

\description{ 
Adattato da Lionetti & al, 2015. Dati relativi ad una ricerca sul ruolo dell'attaccamento e dell'alleanza tra genitori come predittori dello stress in genitori adottivi, viene intervistato un gruppo di 40 coppie con un figlio adottato. Lo stress dei genitori è valutato su tre domini misurati con tre strumenti (Parent Distress, Parent/Child Disfunctional interaction e Difficulty Child).} 

\usage{data( parenting )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{PD.m}: int, punteggi delle madri sul Parent Distress.
    \item \code{PCD.m}: int, punteggi delle madri sul Parent/Child Disfunctional interaction.
    \item \code{CD.m}: int, punteggi delle madri sul Difficulty Child.
    \item \code{PD.p}: int, punteggi dei padri sul Parent Distress.
    \item \code{PCD.p}: int, punteggi dei padri sul Parent/Child Disfunctional interaction.
    \item \code{CD.p}: int, punteggi dei padri sul Difficulty Child.
    \item \code{U.m}: int, attaccamento irrisolto madri (\code{U.m = 1}).
    \item \code{PAM.m}: int, punteggi delle madri sul Parenting Alliance Measure.
    \item \code{U.p}: int, attaccamento irrisolto padri (\code{U.p = 1}).
    \item \code{PAM.p}: int, punteggi dei padri sul Parenting Alliance Measure.
}
}
\references{Lionetti, F., Pastore, M., Barone, L. (2015). Parenting stress: the roles of attachment and parenting alliance in the context of adoption. Parenting: \emph{Science and Practice}, 15, 75--91.}
\author{Massimiliano Pastore}
\keyword{data}
