% --- Source file: pre_post_memory.Rd --- 
\name{pre_post_memory} 
\alias{pre_post_memory} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: simulazione di un esperimento con due gruppi di soggetti cui vengono applicate due diverse tecniche per il miglioramento delle abilità di memoria.} 

\usage{data( pre_post_memory )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{ID}: Factor, codice identificativo.
    \item \code{pre}: num, punteggio di memoria prima del training.
    \item \code{post}: num, punteggio di memoria dopo il training.
    \item \code{tecnica}: Factor, tecnica di miglioramento della memoria utilizzata nel training.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
