% --- Source file: radar.Rd --- 
\name{radar} 
\alias{radar} 
\title{Data set fittizio} 
\description{ 
Data set 
fittizio: esperimento sull'effetto della deprivazione da sonno in un compito attentivo.
I soggetti, deprivati di sonno per un certo numero di ore, devono individuare un oggetto in 
movimento su un radar. 
} 

\usage{data( radar )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{subj}: num, codice del soggetto.
    \item \code{hr}: num, ore di deprivazione da sonno.
    \item \code{score}: num, numero di errori commessi.}
}

\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, London.}
\author{Massimiliano Pastore}
\keyword{data}
