% --- Source file: redditi.Rd ---
\name{redditi}
\alias{redditi}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}
\description{
Data set fittizio: redditi giornalieri di 2000 soggetti di nazionalita' differenti.
}
\usage{
data( redditi )
}
%- maybe also 'usage' for other objects documented here.

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{nazion}: Factor, nazionalita' del soggetto.
    \item \code{reddito}: num, reddito in dollari al giorno.}
}

\author{Massimiliano Pastore}
\keyword{data}
