% --- Source file: relbambini.Rd --- 
\name{relbambini} 
\alias{relbambini} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
punteggi di competenza linguistica in un campione di 81 soggetti.} 

\usage{data( relbambini )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{cat.comp}: Factor, categoria di comportamento.
    \item \code{comp.ling}: num, punteggio di competenza linguistica.
}
}
%\source{Keppel, G. (1991). \emph{Design and analysis}. Prentice Hall, %London.}
\author{Massimiliano Pastore}
\keyword{data}
