% --- Source file: rinforzo.Rd --- 
\name{rinforzo} 
\alias{rinforzo} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
punteggi ad una serie di problemi di ragionamento di un campione di bambini cui vengono dati tre tipi di rinforzo.} 

\usage{data( rinforzo )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{sogg}: int, codice soggetto.
    \item \code{rinforzo}: Factor, tipo di rinforzo.
    \item \code{risp}: int, punteggi.}
}
\source{Keppel, G., Saufley, W. H., Tokunaga, H. (2001). \emph{Disegno sperimentale e analisi dei dati in psicologia}. EdiSes, Napoli.
}
\author{Massimiliano Pastore}
\keyword{data}
