% --- Source file: school.Rd ---
\name{school}
\alias{school}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}

\description{
Data set fittizio: Dati relativi a 260 studenti di scuola superiore.
}

\usage{
data( school )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{schid}: Factor, codice identificativo della scuola.
    \item \code{public}: Factor, status della scuola, pubblica (\code{yes}) o privata (\code{no}).
    \item \code{ratio}: num, rapporto studenti-docenti della scuola.
    \item \code{percmin}: num, percentuale di studenti appartenenti a minoranze etniche.
    \item \code{classid}: Factor, codice identificativo della classe.
    \item \code{stuid}: Factor, codice identificativo dello studente.    
    \item \code{sex}: Factor, genere dello studente.
    \item \code{ethn}: Factor, gruppo etnico dello studente.
    \item \code{ses}: num, status socio-economico della famiglia dello studente.
    \item \code{homework}: num, ore dedicate alla settimana ai compiti di matematica.
    \item \code{math}: num, punteggio al test di matematica.
    }
}
%\source{Tabacknick, B.G., Fidell, L. (1996), \emph{Using Multivariate Statistics}, Harper Collins, NY.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
