% --- Source file: social.Rd ---
\name{social}
\alias{social}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}

\description{
Data set fittizio: ricerca sul clima sociale di quartiere.
}

\usage{
data( social )
}

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{social.capital}: num, misura del supporto tra vicini ed il clima sociale.
    \item \code{safety.concerns}: num, misura della paura della criminalità e
preoccupazione per il proprio figlio.
    \item \code{parenting}: num, misura del grado di sostegno dei genitori.
    \item \code{antisocial.behavior}: num, comportamento antisociale.
    }
}
%\source{Tabacknick, B.G., Fidell, L. (1996), \emph{Using Multivariate Statistics}, Harper Collins, NY.}
\author{Massimiliano Pastore}
%}
\keyword{data}
%\eof
