% --- Source file: sordomuti.Rd ---
\name{sordomuti}
\alias{sordomuti}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data set fittizio}
\description{
Data set fittizio: tempi di esecuzione di una 
prova ad incastro per un campione di 64 soggetti sordomuti.
}
\usage{
data( sordomuti )
}
%- maybe also 'usage' for other objects documented here.

\format{
Il data frame contiene una sola variabile:
    \code{tempo}: num, tempi di esecuzione in secondi.
}

\author{Massimiliano Pastore}
\keyword{data}
