% --- Source file: tmt.Rd --- 
\name{tmt} 
\alias{tmt} 
\title{Data set fittizio} 
\description{ 
Data set 
fittizio: valutazioni al Trail Making Test per 13 soggetti con deficit attentivo.
} 

\usage{data( tmt )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{subj}: num, codice del soggetto.
    \item \code{group}: Factor, gruppo di appartenenza del soggetto: 
            \code{t} = gruppo trattamento, \code{c} = gruppo di controllo.
    \item \code{tmtA1}: num, punteggio al Trail Making Test (forma A) nella prima somministrazione.
    \item \code{tmtA2}: num, punteggio al Trail Making Test (forma A) nella seconda somministrazione.
    \item \code{tmtA3}: num, punteggio al Trail Making Test (forma A) nella terza somministrazione.
    \item \code{tmtB1}: num, punteggio al Trail Making Test (forma B) nella prima somministrazione.
    \item \code{tmtB2}: num, punteggio al Trail Making Test (forma B) nella seconda somministrazione.
    \item \code{tmtB3}: num, punteggio al Trail Making Test (forma B) nella terza somministrazione.}
}

\author{Massimiliano Pastore}
\keyword{data}
