% --- Source file: vaes2015.Rd --- 
\name{vaes2015} 
\alias{vaes2015} 
\title{Data set fittizio} 

\description{ 
Data set fittizio: 
dati relativi ad un campione di 200 soggetti coinvolti in una ricerca sul pregiudizio verso gli immigrati.} 

\usage{data( vaes2015 )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{PREGIUDIZIO}: num, livello di pregiudizio verso gli extracomunitari, più alti sono i punteggi più è alto il livello di pregiudizio.
    \item \code{NORME}: num, livello di accordo espresso dai partecipanti in relazione alla presenza di norme anti discriminazione (es. Le persone che esprimono atteggiamenti offensivi verso gli immigrati devono essere perseguite legalmente).
    \item \code{PAURACRI}: num, livello di paura della criminalità percepita.
    \item \code{STIME}: num, stima soggettiva della percentuale di immigrati che commettono crimini.
    \item \code{CONTATTO}: num, livello di contatto percepito con gli immigrati.
    \item \code{AGENTI_SOC}: num, livello di pregiudizio delle persone importanti (partner, genitori, amici stretti, etc.).
    \item \code{Or_Politico}: num, orientamento politico, espresso su una scala da 1 (estrema sinistra) a 16 (estrema destra).
    \item \code{GIORNALIsx}: num, frequenza di lettura giornali con orientamento di sinistra.
    \item \code{GIORNALIdx}: num, frequenza di lettura giornali con orientamento di destra.
    \item \code{TG}: num, frequenza di esposizione ai telegiornali.
}
}
\references{Vaes, J., Latrofa, M., Vieno, A., Pastore, M. (2015). Exposure to politicized media and prejudice against immigrants in Italy: Identifying its impact and psychological mediators. \emph{Psicologia Sociale}, 10, 141-160.}
\author{Massimiliano Pastore}
\keyword{data}
