% --- Source file: Worland.Rd --- 
\name{Worland} 
\alias{Worland} 
\title{Data set} 

\description{ 
Data set: 
dati relativi ad una ricerca .} 

\usage{data( Worland )} 

\format{
Il data frame contiene le seguenti variabili:
    \itemize{
    \item \code{parpsych}: num, 
    \item \code{lowses}: num, 
    \item \code{verbal}: num, 
    \item \code{vissspa}: num, 
    \item \code{memory}: num, 
    \item \code{reading}: num, 
    \item \code{arithm}: num, 
    \item \code{spelling}: num, 
    \item \code{motiv}: num, motivazione scolastica. 
    \item \code{extrav}: num, estroversione.
    \item \code{harmony}: num, armonia.
    \item \code{stability}: num, stabilità emotiva.
}
}
\source{Worland, J., Weeks, D. G., Janes, C. L., & Strock, B. D. (1984). Intelligence, classroom behavior, and academic achievement in children at high and low risk for psychopathology: A structural equation analysis. \emph{Journal of abnormal child psychology}, 12, 437--454.}
\author{Massimiliano Pastore}
\keyword{data}
