\name{ADati-package}
\alias{ADati-package}
%- Also NEED an '\alias' for EACH other topic documented here.
%\inputencoding{utf8}
%\usepackage[utf8]{inputenc}

\title{ADati package}
\description{
Contiene una miscellanea di funzioni associate al manuale Analisi dei dati quantitativi in Psicologia.

NOTA: Alcune funzioni richiedono il pacchetto \code{gtools} e il pacchetto \code{MASS}.
}

\details{
\tabular{ll}{
Package: \tab ADati\cr
Type: \tab Package\cr
Version: \tab 5.5.3\cr
Depends: \tab R (>= 3.5.0)\cr
Imports: \tab ggplot2, lavaan, reshape, heplots, rstanarm\cr
Suggests: \tab MASS, gtools, brms\cr
Date: \tab 2025-03-07\cr
License: \tab GPL-2\cr
Encoding: \tab UTF-8\cr
}
}
\author{Massimiliano Pastore}
\references{
Pastore, M. (2015). Analisi dei dati in Psicologia (e applicazioni con R). Il Mulino, Bologna. URL: \url{https://lilia.dpss.psy.unipd.it/~massimiliano.pastore/Ilmulino.html}.
}

